+++ app/components/H2/index.js
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import { styled } from 'linaria/react';
 
 const H2 = styled.h2`
   font-size: 1.5em;