+++ app/components/Header/HeaderLink.js
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
+import { styled } from 'linaria/react';
 
 export default styled(Link)`
   display: inline-flex;