+++ app/components/List/Ul.js
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import { styled } from 'linaria/react';
 
 const Ul = styled.ul`
   list-style: none;