+++ app/containers/RepoListItem/tests/IssueIcon.test.js
@@ -1,16 +1,10 @@
 import React from 'react';
 import { render } from 'react-testing-library';
-import renderer from 'react-test-renderer';
-import 'jest-styled-components';
+import 'linaria-jest';
 
 import IssueIcon from '../IssueIcon';
 
 describe('<IssueIcon />', () => {
-  it('should match the snapshot', () => {
-    const renderedComponent = renderer.create(<IssueIcon />).toJSON();
-    expect(renderedComponent).toMatchSnapshot();
-  });
-
   it('should have a className attribute', () => {
     const { container } = render(<IssueIcon />);
     expect(container.firstChild.hasAttribute('class')).toBe(true);