+++ app/containers/RepoListItem/tests/RepoLink.test.js
@@ -1,16 +1,10 @@
 import React from 'react';
 import { render } from 'react-testing-library';
-import renderer from 'react-test-renderer';
-import 'jest-styled-components';
+import 'linaria-jest';
 
 import RepoLink from '../RepoLink';
 
 describe('<RepoLink />', () => {
-  it('should match the snapshot', () => {
-    const renderedComponent = renderer.create(<RepoLink />).toJSON();
-    expect(renderedComponent).toMatchSnapshot();
-  });
-
   it('should have a className attribute', () => {
     const { container } = render(<RepoLink />);
     expect(container.firstChild.hasAttribute('class')).toBe(true);