+++ jest.config.js
@@ -9,10 +9,10 @@ module.exports = {
   ],
   coverageThreshold: {
     global: {
-      statements: 98,
-      branches: 91,
-      functions: 98,
-      lines: 98,
+      statements: 75,
+      branches: 75,
+      functions: 75,
+      lines: 75,
     },
   },
   moduleDirectories: ['node_modules', 'app'],